pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ibkr-telegram"
version = "0.1.0"
description = "Telegram bot for Interactive Brokers: portfolio management, order execution, safety checks"
readme = "README.md"
license = "GPL-3.0-only"
requires-python = ">=3.12"
authors = [{ name = "GeiserX" }]

dependencies = [
    # Telegram
    "aiogram>=3.15",

    # IBKR
    "ib-async>=1.0.3",

    # Core
    "pyyaml>=6.0",
    "aiosqlite>=0.20",
    "defusedxml>=0.7",

    # Docker control (gateway pause/resume)
    "docker>=7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "ruff>=0.8",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
